fix 1

+++ index1.js
@@ -139,8 +139,15 @@ async function readFilesFromFlexy(){
 				  .on('end', async function(){
             //console.log(arrData)
             if (arrData.length == 0) {
-              fs.copyFileSync(currentPath, errPath);
-              fs.unlinkSync(currentPath)
+              try{
+                if( fs.existsSync(currentPath) ){
+                  fs.copyFileSync(currentPath, errPath);
+                  fs.unlinkSync(currentPath)
+                }
+              }catch(err){
+                console.log(err.message)
+              }
+              
             }else{
               let sts = await SaveDataToSQLServer(arrData)
               console.log('SQL', site_id,':',sts)
@@ -153,9 +160,16 @@ async function readFilesFromFlexy(){
                 let strPathFile = _strPath_Date + '\\' + moment(new Date()).format("YYYYMMDD-HHmmss") + '_' + file
 
                 const folderDate = mkdirp.sync(_strPath_Date);
-
-                fs.copyFileSync(currentPath, strPathFile);
-                fs.unlinkSync(currentPath)
+                try{
+                  if( fs.existsSync(currentPath) ){
+                    fs.copyFileSync(currentPath, strPathFile);
+                    fs.unlinkSync(currentPath)
+                  }
+                  
+                }catch(err){
+                  console.log('Error delete file: '+ err.message)
+                }
+                
               }
               sendAckToFlexy(site_id, ackTag, ip, port);
             }
